WIP: drm/msm: extend the submit ioctl to pass in flags
We'll want to be able to pass in flags, such as asking for explicit
fencing, and possibly other things down the road.  Fortunately we don't
need a full 32b for the pipe-id.  So use the upper 16 bits (which could
be extended later if needed) for flags.  Since anything with the upper
bits set would not be a valid pipe-id, an old userspace would not set
any of the upper bits, and an old kernel would reject it as an invalid
pipe-id.