DOWNSTREAM: drm/msm/mdp5: Get the correct platform device for bus scaling
With the new DT hierarchy in place, the bus scaling properties are to be
parsed from the device node corresponding to the MDP5 device. The existing
code tries to retrieve it from the device node corresponding to the MDSS
device.

Retrieve the correct platform device pointer (the one corresponding to
MDP5) and pass that to the bus scaling API.

Signed-off-by: Archit Taneja <architt@codeaurora.org>