HACK: drm/msm: increase coherent dma mask
Increase the coherent DMA mask to a higher number. This is the
physical memory that the IOMMU maps to a 32 bit VA for display.
The physical memory chunks provided by shmem come out to be 33
bit on 8996, and the existing mask causes the driver to reject
these chunks.

This is a short term fix. Need to find a better way to set the
correct mask.

Signed-off-by: Archit Taneja <architt@codeaurora.org>