drm: move check for min/max width/height for atomic drivers
drm-hwc + android tries to create an fb for the wallpaper layer, which
is larger than the screen resolution, and potentially larger than
mode_config->max_{width,height}.  But the plane src_w/src_h is within
the max limits, so it is something the hardware can otherwise do.

For atomic drivers, defer the check to drm_atomic_plane_check().

Signed-off-by: Rob Clark <robdclark@gmail.com>