drm: make drm_get_format_name atomic/irq safe again
Fixes: 90844f00049e9f42573fd31d7c32e8fd31d3fd07

    drm: make drm_get_format_name thread-safe

    Signed-off-by: Eric Engestrom <eric@engestrom.ch>
    [danvet: Clarify that the returned pointer must be freed with
    kfree().]
    Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>

Note: I think we need to be a bit careful about follow-up audits of
callers of this..  now that you need to kfree the return value I think
it is fairly easy for new patches to introduce new callers which leak
the return value.  We probably should have left the function as-is and
introduce a new variant, or something like that.

Signed-off-by: Rob Clark <robdclark@gmail.com>